build-faster-llm-as-a-service-for-developers-with-qwen-3-5-122b/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytest>=8.0.0
